requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi @ file:///C:/b/abs_78eb1_vq6z/croot/cffi_1714483206096/work
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.6
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
frozenlist==1.4.1
fsspec @ file:///C:/b/abs_d5jzmndkba/croot/fsspec_1714461594761/work
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///C:/ci/gmpy2_1645455782955/work
greenlet==3.0.3
idna==3.7
Jinja2 @ file:///C:/b/abs_92fccttino/croot/jinja2_1716993447201/work
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.5
langchain-text-splitters==0.2.1
langsmith==0.1.75
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/b/abs_ecfdqh67b_/croot/markupsafe_1704206030535/work
marshmallow==3.21.3
mdurl==0.1.2
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
more-itertools==10.2.0
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
multidict==6.0.5
mypy-extensions==1.0.0
networkx @ file:///C:/b/abs_3bxnu56g9d/croot/networkx_1717597507456/work
numba==0.59.1
numpy @ file:///C:/b/abs_c1ywpu18ar/croot/numpy_and_numpy_base_1708638681471/work/dist/numpy-1.26.4-cp39-cp39-win_amd64.whl#sha256=f284633067fde4bb078a09085a1c9e99a3218c129f5312b7b3d76c9340442a25
openai-whisper==20231117
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pyarrow==16.1.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.11
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.30
streamlit==1.35.0
sympy @ file:///C:/b/abs_82njkonm7f/croot/sympy_1701397685028/work
tenacity==8.3.0
tiktoken==0.7.0
toml==0.10.2
toolz==0.12.1
torch @ file:///C:/b/abs_6fueooay2f/croot/pytorch-select_1707342446212/work
tornado==6.4.1
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions @ file:///C:/b/abs_0as9mdbkfl/croot/typing_extensions_1715268906610/work
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.1
yarl==1.9.4
